refactor: rename main package name from io.github.mdaubie.srtparser to io.github.mdaubie.subtitlesparser
