+++ src/main/java/io/github/mdaubie/subtitlesparser/model/Format.java
@@ -1,4 +1,4 @@
-package io.github.mdaubie.srtparser.model;
+package io.github.mdaubie.subtitlesparser.model;
 
 import java.time.format.DateTimeFormatter;
