+++ src/test/java/io/github/mdaubie/subtitlesparser/ParserTest.java
@@ -1,9 +1,9 @@
-package io.github.mdaubie.srtparser;
+package io.github.mdaubie.subtitlesparser;
 
-import io.github.mdaubie.srtparser.model.Format;
-import io.github.mdaubie.srtparser.model.SubRipFile;
-import io.github.mdaubie.srtparser.model.SubRipSubtitle;
-import io.github.mdaubie.srtparser.model.SubtitlesFile;
+import io.github.mdaubie.subtitlesparser.model.Format;
+import io.github.mdaubie.subtitlesparser.model.SubRipFile;
+import io.github.mdaubie.subtitlesparser.model.SubRipSubtitle;
+import io.github.mdaubie.subtitlesparser.model.SubtitlesFile;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
